requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.4
alembic==1.13.2
cryptography==42.0.8
httpx==0.27.0
python-multipart==0.0.9
redis==5.0.4
celery==5.4.0
pytest==8.2.2
pytest-asyncio==0.23.7
asyncpg==0.29.0
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyJWT==2.9.0
pyroute2==0.7.12
argon2-cffi==23.1.0
structlog==24.1.0
prometheus-client==0.20.0
grpcio==1.65.1
grpcio-tools==1.65.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
black==24.4.2
ruff==0.5.6
pre-commit==3.7.1
qrcode==7.4.2
